[FE-970] moved OGOnlyWrapper to hibana folder
