+++ src/components/hibana/index.js
@@ -1,5 +1,6 @@
 export { default as HibanaBanner } from './HibanaBanner';
 export { default as HibanaToggle } from './HibanaToggle';
+export { default as OGOnlyWrapper } from './OGOnlyWrapper';
 export { default as HibanaDescription } from './HibanaDescription';
 export { default as HibanaStyleHandler } from './HibanaStyleHandler';
 export { default as Navigation } from './Navigation';
