+++ src/pages/abTesting/components/view/SettingsView.js
@@ -1,5 +1,6 @@
 import React from 'react';
-import { OGOnlyWrapper, Panel } from 'src/components/matchbox';
+import { Panel } from 'src/components/matchbox';
+import { OGOnlyWrapper } from 'src/components/hibana';
 import { LabelledValue } from 'src/components';
 import { snakeToFriendly } from 'src/helpers/string';
